pawpals_api/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
requests==2.31.0 # For Nominatim/OSM API calls if needed

